Split Cartridge into files and some cleanup
