+++ Source/Cartridge/MBC5.py
@@ -1,0 +1,12 @@
+# -*- encoding: utf-8 -*-
+#
+# Authors: Asger Anders Lund Hansen, Mads Ynddal and Troels Ynddal
+# License: See LICENSE file
+# GitHub: https://github.com/Baekalfen/PyBoy
+#
+from GenericMBC import GenericMBC
+
+class MBC5(GenericMBC):
+    def __setitem__(self, address, value):
+        raise Exception("Not implemented")
+
