+++ Source/MB/MemoryManager.py
@@ -1,6 +1,6 @@
 # -*- encoding: utf-8 -*-
 #
-# Authors: Mads Ynddal
+# Authors: Asger Anders Lund Hansen, Mads Ynddal and Troels Ynddal
 # License: See LICENSE file
 # GitHub: https://github.com/Baekalfen/PyBoy
 #
